4_sqldb_tshirts/requirements.txt
langchain==0.1.7
langchain-text-splitters==0.3.5
langchain-cli==0.0.35
langserve==0.0.17
langsmith==0.0.87
python-dotenv==1.0.0
streamlit==1.41.0
tiktoken==0.8.0
faiss==1.9.0
faiss-cpu==1.9.0
protobuf==4.25.6
langchain_experimental==0.0.17
langchain-community==0.0.20
google-generativeai==0.3.2
mysql-connector-python==9.2.0
pymysql==1.1.1
sentence-transformers==3.4.1
chromadb==0.6.3
chroma-hnswlib==0.7.6
google-ai-generativelanguage==0.4.0
google-generativeai==0.3.2
grpcio==1.70.0
grpcio-status==1.63.0rc1
huggingface-hub==0.28.1
openai==1.60.1
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
pydantic==1.10.21
pydantic_core==2.27.2
pydantic-settings==2.6.1
transformers==4.48.2
validators==0.34.0